chore: update TS target to ES2022
With webcontainers now enabled by default, the CLI is building the angular app. The CLI downlevels native async/awaits (via babel) to support zone.js patching and there is no need to target <ES2017 anymore (where TSC was doing the down leveling).

Moving to ES2022 removes the following warning in the console: `TypeScript compiler options "target" and "useDefineForClassFields" are set to "ES2022" and "false" respectively by the Angular CLI.` when starting a project