fix: add proper redirect to tabs
Introduce tab-based navigation with a home screen. The `(tabs)` layout includes a tab bar with a home icon, and the root layout is updated to include the tab screen. The index page redirects to the tab layout on load thus fixing the "This screen doesn't exist" display when loading the app on Expo Go